% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\docType{data}
\name{conf}
\alias{conf}
\title{Parámetros de configuración y mantenimiento}
\format{
An object of class \code{list} of length 7.
}
\usage{
conf
}
\description{
Lista con párametros de configuración y mantenimiento. Incluye patrones
para identificar nombre de bases de datos y urls en htmls, además de el
url con estadísticas del MEFD.
}
\keyword{data}
