% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefd_fun.R
\name{mefd_meta}
\alias{mefd_meta}
\title{metadata de indicadores}
\usage{
mefd_meta(url_web, config = conf)
}
\arguments{
\item{url_web}{url de página web del MEFD con lista de indicadores}

\item{config}{lista con parámetros de configuración (default)}
}
\value{
tibble con metadatos
}
\description{
Genera metadata para indicadores a partir de url del MEFD: nombre de archivos,
urls de archivos, nombre del indicador
}
\examples{
mi_url <- paste0("https://estadisticas.educacion.gob.es/EducaDynPx/",
"educabase/index.htm?type=pcaxis&path=/no-universitaria/centros/",
"centrosyunid/series/unidades_esc&file=pcaxis&l=s0")
mi_meta <- mefd_meta(mi_url)
}
