% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mefd_fun.R
\name{mefd_name}
\alias{mefd_name}
\title{Nombre de indicadores}
\usage{
mefd_name(url_web, config = conf)
}
\arguments{
\item{url_web}{url de página web del MEFD con lista de indicadores}

\item{config}{lista con parámetros de configuración (default)}
}
\value{
vector con el nombre de los archivos .csv
}
\description{
Genera vector con nombre de bases de datos (.csv) para cada indicador
a partir de url de la web
}
\examples{
mi_url <- paste0("https://estadisticas.educacion.gob.es/EducaDynPx/",
"educabase/index.htm?type=pcaxis&path=/no-universitaria/centros/",
"centrosyunid/series/unidades_esc&file=pcaxis&l=s0")
mefd_name(mi_url)
}
