% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\docType{data}
\name{meta_mefd}
\alias{meta_mefd}
\title{meta_mefd}
\format{
An object of class \code{data.frame} with 269 rows and 7 columns.
}
\usage{
meta_mefd
}
\description{
Metadatos de indicadores. Incluye el nombre del indicadore, el archivo .csv, el url.
}
\keyword{datasets}
