% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\docType{data}
\name{par_url}
\alias{par_url}
\title{par_url}
\format{
An object of class \code{data.frame} with 15 rows and 4 columns.
}
\usage{
par_url
}
\description{
Parámetros con url de páginas del MEFD para crear meta_mefd
}
\keyword{datasets}
